jms: Move listener code to common module
This is done to reuse this module in both apim and mgw
