+++ components/apimgt/org.wso2.carbon.apimgt.common.jms/pom.xml
@@ -9,27 +9,30 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
     <packaging>bundle</packaging>
-    <artifactId>org.wso2.carbon.apimgt.jms.listener</artifactId>
+    <artifactId>org.wso2.carbon.apimgt.common.jms</artifactId>
 
     <dependencies>
         <dependency>
-            <groupId>org.wso2.carbon.apimgt</groupId>
-            <artifactId>org.wso2.carbon.apimgt.impl</artifactId>
+            <groupId>org.apache.axis2</groupId>
+            <artifactId>axis2-transport-base</artifactId>
         </dependency>
         <dependency>
-            <groupId>org.apache.felix</groupId>
-            <artifactId>org.apache.felix.scr.ds-annotations</artifactId>
-            <scope>provided</scope>
+            <groupId>javax.jms</groupId>
+            <artifactId>javax.jms-api</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.wso2.andes.wso2</groupId>
+            <artifactId>andes-client</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>commons-logging</groupId>
+            <artifactId>commons-logging</artifactId>
         </dependency>
         <dependency>
             <groupId>org.jacoco</groupId>
             <artifactId>org.jacoco.agent</artifactId>
             <classifier>runtime</classifier>
             <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.wso2.carbon.apimgt</groupId>
-            <artifactId>org.wso2.carbon.apimgt.common.jms</artifactId>
         </dependency>
     </dependencies>
 
@@ -41,27 +44,20 @@
                 <extensions>true</extensions>
                 <configuration>
                     <instructions>
-                        <Bundle-SymbolicName>
-                            ${project.artifactId}
-                        </Bundle-SymbolicName>
+                        <Bundle-SymbolicName>${project.artifactId}</Bundle-SymbolicName>
+                        <Bundle-Name>${project.artifactId}</Bundle-Name>
                         <Private-Package>
-                            org.wso2.carbon.apimgt.jms.listener.internal.*
                         </Private-Package>
+                        <Export-Package>
+                            org.wso2.carbon.apimgt.common.jms.*;version="${carbon.apimgt.version}",
+                        </Export-Package>
                         <Import-Package>
+                            com.google.gson.annotations.*,
                             javax.transaction.*,
                             org.apache.axis2.*,
                             org.apache.commons.*,
-                            org.osgi.service.*;version="${imp.package.version.osgi.service}",
-                            org.wso2.carbon.apimgt.impl.dto.*;version="${carbon.apimgt.imp.pkg.version}",
-                            org.wso2.carbon.apimgt.broker.lifecycle.service.*;version="${carbon.apimgt.imp.pkg.version}",
-                            org.wso2.carbon.context.*;version="${carbon.kernel.package.import.version.range}",
-                            org.wso2.carbon.utils.*;version="${carbon.kernel.package.import.version.range}",
                             *;resolution:=optional
                         </Import-Package>
-                        <Export-Package>
-                            !org.wso2.carbon.apimgt.jms.listener.internal.*,
-                            org.wso2.carbon.apimgt.jms.listener.*;version="${carbon.apimgt.version}"
-                        </Export-Package>
                     </instructions>
                 </configuration>
             </plugin>
