+++ components/apimgt/org.wso2.carbon.apimgt.common.jms/src/main/java/org/wso2/carbon/apimgt/common/jms/ListenerConstants.java
@@ -1,0 +1,31 @@
+/*
+ * Copyright (c) 2021, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.wso2.carbon.apimgt.common.jms;
+
+/**
+ * Used for keeping constants used in JMS Listener
+ */
+public class ListenerConstants {
+
+    public static final String ANDES_SYMBOLIC_NAME = "org.wso2.carbon.andes";
+    public static final String QPID_SERVICE = "org.wso2.carbon.andes.service.QpidService";
+
+    public static final String MB_PROPERTIES = "mb.properties";
+    public static final String CONNECTION_FACTORY_NAME = "Siddhi-JMS-Consumer";
+}
