+++ components/apimgt/org.wso2.carbon.apimgt.common.jms/src/main/java/org/wso2/carbon/apimgt/common/jms/factory/JMSTaskManagerFactory.java
@@ -1,27 +1,29 @@
 /*
-*  Copyright (c) 2016, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
-*
-*  WSO2 Inc. licenses this file to you under the Apache License,
-*  Version 2.0 (the "License"); you may not use this file except
-*  in compliance with the License.
-*  You may obtain a copy of the License at
-*
-*    http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing,
-* software distributed under the License is distributed on an
-* "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-* KIND, either express or implied.  See the License for the
-* specific language governing permissions and limitations
-* under the License.
-*/
+ * Copyright (c) 2021, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
 
-package org.wso2.carbon.apimgt.jms.listener.utils;
+package org.wso2.carbon.apimgt.common.jms.factory;
 
 import org.apache.axis2.description.Parameter;
 import org.apache.axis2.transport.base.BaseConstants;
 import org.apache.axis2.transport.base.threads.WorkerPool;
-import org.wso2.carbon.apimgt.jms.listener.ThrottlingRunTimeException;
+import org.wso2.carbon.apimgt.common.jms.JMSConstants;
+import org.wso2.carbon.apimgt.common.jms.JMSTaskManager;
+import org.wso2.carbon.apimgt.common.jms.JmsRunTimeException;
 
 import javax.jms.Session;
 import java.util.HashMap;
@@ -180,7 +182,7 @@ public class JMSTaskManagerFactory {
             value = cfMap.get(key);
         }
         if (value == null) {
-            throw new ThrottlingRunTimeException("Service/connection factory property : " + key);
+            throw new JmsRunTimeException("Service/connection factory property : " + key);
         }
         return value;
     }
@@ -238,7 +240,7 @@ public class JMSTaskManagerFactory {
             try {
                 return Integer.parseInt(value);
             } catch (NumberFormatException e) {
-                throw new ThrottlingRunTimeException("Invalid value : " + value + " for " + key);
+                throw new JmsRunTimeException("Invalid value : " + value + " for " + key);
             }
         }
         return null;
@@ -261,7 +263,7 @@ public class JMSTaskManagerFactory {
             try {
                 return Double.parseDouble(value);
             } catch (NumberFormatException e) {
-                throw new ThrottlingRunTimeException("Invalid value : " + value + " for " + key);
+                throw new JmsRunTimeException("Invalid value : " + value + " for " + key);
             }
         }
         return null;
@@ -289,7 +291,7 @@ public class JMSTaskManagerFactory {
             } else if (BaseConstants.STR_TRANSACTION_LOCAL.equalsIgnoreCase(val)) {
                 return BaseConstants.TRANSACTION_LOCAL;
             } else {
-                throw new ThrottlingRunTimeException("Invalid option : " + val + " for parameter : " + BaseConstants.STR_TRANSACTION_JTA);
+                throw new JmsRunTimeException("Invalid option : " + val + " for parameter : " + BaseConstants.STR_TRANSACTION_JTA);
                 //TODO fix it
                 //return 0;
             }
@@ -344,7 +346,7 @@ public class JMSTaskManagerFactory {
             try {
                 return Integer.parseInt(val);
             } catch (NumberFormatException ignore) {
-                throw new ThrottlingRunTimeException("Invalid session acknowledgement mode : " + val);
+                throw new JmsRunTimeException("Invalid session acknowledgement mode : " + val);
                 //TODO remove this
                 //return 0;
             }
@@ -375,7 +377,7 @@ public class JMSTaskManagerFactory {
         } else if ("consumer".equals(val)) {
             return JMSConstants.CACHE_CONSUMER;
         } else if (val != null) {
-            throw new ThrottlingRunTimeException("Invalid cache level : " + val);
+            throw new JmsRunTimeException("Invalid cache level : " + val);
         }
         return JMSConstants.CACHE_AUTO;
     }
