+++ components/apimgt/org.wso2.carbon.apimgt.gateway/pom.xml
@@ -112,7 +112,7 @@
         </dependency>
         <dependency>
             <groupId>org.wso2.carbon.apimgt</groupId>
-            <artifactId>org.wso2.carbon.apimgt.jms.listener</artifactId>
+            <artifactId>org.wso2.carbon.apimgt.common.jms</artifactId>
         </dependency>
         <dependency>
             <groupId>org.wso2.carbon</groupId>
@@ -455,6 +455,7 @@
                             com.nimbusds.jwt.*,
                             org.apache.commons.codec.*,
                             org.wso2.carbon.apimgt.common.gateway.*,
+                            org.wso2.carbon.apimgt.common.jms.*,
                             *;resolution:=optional
                         </Import-Package>
                     </instructions>
