+++ components/apimgt/pom.xml
@@ -59,5 +59,6 @@
         <module>org.wso2.carbon.apimgt.rest.api.common</module>
         <module>org.wso2.carbon.apimgt.rest.api.publisher.v1.common</module>
         <module>org.wso2.carbon.apimgt.common.gateway</module>
+        <module>org.wso2.carbon.apimgt.common.jms</module>
     </modules>
 </project>
