+++ features/apimgt/org.wso2.carbon.apimgt.jms.listener.feature/pom.xml
@@ -36,6 +36,10 @@
             <groupId>org.wso2.carbon.apimgt</groupId>
             <artifactId>org.wso2.carbon.apimgt.jms.listener</artifactId>
         </dependency>
+        <dependency>
+            <groupId>org.wso2.carbon.apimgt</groupId>
+            <artifactId>org.wso2.carbon.apimgt.common.jms</artifactId>
+        </dependency>
     </dependencies>
 
     <build>
@@ -65,6 +69,9 @@
                                 </bundleDef>
                                 <bundleDef>
                                     org.wso2.carbon.apimgt:org.wso2.carbon.apimgt.broker.lifecycle:${carbon.apimgt.version}
+                                </bundleDef>
+                                <bundleDef>
+                                    org.wso2.carbon.apimgt:org.wso2.carbon.apimgt.common.jms:${carbon.apimgt.version}
                                 </bundleDef>
                             </bundles>
                         </configuration>
